requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.1
requests
backoff
pandas
python-dotenv
py2app
wheel
setuptools
chardet
charset_normalizer
tqdm
selenium==4.1.0
aiohttp
aiosqlite